# wfzo-api-main/wfzo-api-main/api-spec.yml
openapi: 3.0.4
info:
  title: WFZO Website Management Platform API
  description: >-
    REST APIs for the World Free Zones Organization (WFZO) web platform.
    Scopes
     - Auth: Microsoft Entra (External ID) Bearer JWT
     - Content: read-through to Strapi CMS
     - Documents: secure PDF delivery via short-lived Azure Blob SAS
     - Events: Zoom registration with per-membership seat quotas
     - Learning: LMS integration with Moodle
     - Community: Discourse integrations (DMs, topics, embeds)
     - CMS APIs
  version: 1.0.0
servers:
  - url: /wfzo/api/v1
...
paths:
...
  /document/upload:
    post:
      tags:
        - Document Management
      summary: Upload a new document (multipart)
      description:
        Upload a file with optional metadata. For large files, chunked uploads can be handled out-of-band; this endpoint
        registers and stores the final artifact. Usecases like
        - Member Logo uploads
        - Hero video Streaming in the home page(video uploaded via Strapi cloud to BLOB storage, but the streaming will be done via
        backend services)
        - FAQ videos also the same route
        - in general any PDF, Video, Image uploads to handle from Website/admin portal
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadForm"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAsset"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /document/{id}/download:
    get:
      tags:
        - Document Management
      summary: Get a time-limited download URL
      description: Returns a signed URL object for downloading or streaming the requested variant.
        - Member Logo dowloads
        - Hero video Streaming in the home page(video uploaded via Strapi cloud to BLOB storage, but the streaming will be done via
        backend services)
        - FAQ videos also the same route
        - in general any PDF, Video, Image downloads to handle from Website/admin portal
      parameters:
        - name: id
          in: path
          description: Document identifier
          required: true
          schema:
            type: string
        - name: variant
          in: query
          description: Requested asset variant
          required: false
          schema:
            type: string
            enum: [original, thumbnail, preview, hlsManifest]
        - name: inline
          in: query
          description: Request inline Content-Disposition where possible
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignedUrlResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /document/{id}/status:
    get:
      tags:
        - Document Management
      summary: Get document processing status
      description: Returns processing status and available variants for the given document.
      parameters:
        - name: id
          in: path
          description: Document identifier
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
components:
...
  schemas:
    MediaKind:
      type: string
      description: Logical media type classification
      enum: [document, image, video]
    DocumentUploadForm:
      type: object
      description: Multipart form to upload any asset (image, video, document)
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: The file to upload
        mediaKind:
          $ref: "#/components/schemas/MediaKind"
        purpose:
          type: string
          description: Logical usage context (e.g., member-logo, event-hero, library-asset)
        checksum:
          type: string
          description: Optional SHA-256 checksum (hex) for integrity verification
        fileName:
          type: string
          description: Optional explicit file name override
        contentType:
          type: string
          description: Optional explicit MIME type; server may override based on detection
    DocumentVariant:
      type: object
      description: A concrete downloadable/streamable rendition of the asset
      required: [key, url, contentType]
      properties:
        key:
          type: string
          enum: [original, thumbnail, preview, hlsManifest]
        url:
          type: string
          format: uri
        contentType:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        bitrateKbps:
          type: integer
          nullable: true
        size:
          type: integer
          nullable: true
        ready:
          type: boolean
          description: Whether this variant is ready to be consumed
          default: true
    DocumentAsset:
      type: object
      description: Stored asset with metadata and available variants
      required: [id, fileName, contentType, size, mediaKind, createdAt]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        contentType:
          type: string
        size:
          type: integer
          description: Size in bytes
        mediaKind:
          $ref: "#/components/schemas/MediaKind"
        createdAt:
          type: string
          format: date-time
        variants:
          type: array
          items:
            $ref: "#/components/schemas/DocumentVariant"
    SignedUrlResponse:
      type: object
      description: A time-limited URL to download/stream the asset
      required: [url, expiresAt]
      properties:
        url:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
        contentType:
          type: string
        fileName:
          type: string
        size:
          type: integer
          nullable: true
        disposition:
          type: string
          enum: [inline, attachment]
          default: attachment
        variant:
          type: string
          enum: [original, thumbnail, preview, hlsManifest]
          nullable: true
    DocumentStatus:
      type: object
      description: Processing status and variants for a document/asset
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [processing, ready, failed]
        message:
          type: string
          nullable: true
        variants:
          type: array
          items:
            $ref: "#/components/schemas/DocumentVariant"
        updatedAt:
          type: string
          format: date-time
...
    ApiResponse:
      type: object
      description: Standard API response envelope for status reporting
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: "##default"
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header